+++ gittest/osops.py
@@ -2,18 +2,18 @@
 A simple module for common OS Operations
 """
 
-from os.path import basename, dirname
+from pathlib import Path
 
 
 def get_filename(path: str) -> str:
     """
     Get the filename from `path`
     """
-    return basename(path)
+    return Path(path).name
 
 
 def get_foldername(path: str) -> str:
     """
     Get the last foldername from `path`
     """
-    return basename(dirname(path))
+    return Path(path).parent.name