+++ CHANGELOG.md
@@ -1,14 +1,22 @@
 # Changelog
 
-## 2021-04-29 23:27 PM - master
+## 2021-04-29 23:27 - master
 
 - COMMIT: First commit
 
-## 2021-04-29 23:41 PM - feature/folder
+## 2021-04-29 23:41 - feature/folder
 
 - COMMIT: Add function get_foldername in osops file
 
-## 2021-04-29 23:43 PM - master
+## 2021-04-29 23:43 - master
 
 - PR: Merge feature/folder with get_foldername to master
 - Since its just a addition, merged automatically
+
+## 2021-04-30 00:04 - feature/change-to-pathlib
+
+- COMMIT: Change osops to use pathlib instead of os.path
+
+## 2021-04-30 00:08 - feature/change-to-pathlib
+
+- COMMIT: Add pathlib to .vscode/settings.json.cSpell